+++ app/layout.tsx
@@ -4,6 +4,7 @@ import './globals.css'
 import { NavigationMenuPage } from '@/components/navbar/NavigationMenuPage'
 import { Footer } from '@/components/footer/Footer'
 import { ClientProviders } from '@/components/layout/ClientProviders'
+import { DataProvider } from '@/components/layout/DataProvider'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -57,11 +58,13 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
       <body className={inter.className}>
         <div className='flex flex-col min-h-screen'>
           <ClientProviders>
-            <div className='h-16 md:h-16'>
-              <NavigationMenuPage />
-            </div>
-            <main className='flex-grow flex justify-center items-center'>{children}</main>
-            <Footer />
+            <DataProvider>
+              <div className='h-16 md:h-16'>
+                <NavigationMenuPage />
+              </div>
+              <main className='flex-grow flex justify-center items-center'>{children}</main>
+              <Footer />
+            </DataProvider>
           </ClientProviders>
         </div>
       </body>
